+++ openstack/rds/v3/configurations/Delete.go
@@ -1,0 +1,13 @@
+package configurations
+
+import (
+	"github.com/opentelekomcloud/gophertelekomcloud"
+	"github.com/opentelekomcloud/gophertelekomcloud/openstack"
+)
+
+// Delete will permanently delete a particular Configuration based on its unique ID.
+func Delete(c *golangsdk.ServiceClient, id string) (r DeleteResult) {
+	reqOpt := &golangsdk.RequestOpts{OkCodes: []int{200}, MoreHeaders: openstack.StdRequestOpts().MoreHeaders}
+	_, r.Err = c.Delete(c.ServiceURL("configurations", id), reqOpt)
+	return
+}
