+++ openstack/rds/v3/configurations/Get.go
@@ -1,0 +1,12 @@
+package configurations
+
+import (
+	"github.com/opentelekomcloud/gophertelekomcloud"
+	"github.com/opentelekomcloud/gophertelekomcloud/openstack"
+)
+
+// Get retrieves a particular Configuration based on its unique ID.
+func Get(c *golangsdk.ServiceClient, id string) (r GetResult) {
+	_, r.Err = c.Get(c.ServiceURL("configurations", id), &r.Body, openstack.StdRequestOpts())
+	return
+}
